refactor#12-Prifole 경로수정
